Initial refactor test
